Move layout components and respective css
