+++ src/app/(site)/layout.tsx
@@ -3,14 +3,14 @@ import dynamic from 'next/dynamic';
 import clsx from 'clsx';
 
 import Footer from '@/components/Footer';
-import MainMenu from '@/components/MainMenu';
-import MainLogo from '@/components/MainLogo';
-import PathMarker from '@/components/PathMarker';
+import MainMenu from '@/components/layout/MainMenu';
+import MainLogo from '@/components/layout/MainLogo';
+import PathMarker from '@/components/layout/PathMarker';
 
 import '@/styles/globals.scss';
 import styles from './layout.module.scss';
 
-const Background = dynamic(() => import('@/components/Background'), {
+const Background = dynamic(() => import('@/components/layout/Background'), {
   ssr: false,
 });
