+++ src/components/layout/MainLogo.tsx
@@ -1,11 +1,9 @@
 import Image from 'next/image';
 import Link from 'next/link';
 
-import styles from './mainlogo.module.scss';
-
 export default function MainLogo() {
   return (
-    <Link href="/" className={styles.logo}>
+    <Link href="/" className="main-logo">
       <Image src="/images/logo.jpg" width={156} height={156} alt="" />
       <span>Black Dog Blues Band</span>
     </Link>
