+++ src/styles/layout/main-logo.scss
@@ -1,6 +1,6 @@
 @use '@/styles/breakpoints.scss' as breakpoints;
 
-.logo {
+.main-logo {
   display: flex;
   flex-direction: column;
   gap: 1rem;
