+++ src/styles/layout/main-menu.scss
@@ -1,4 +1,4 @@
-.menu {
+.main-menu {
   ul {
     margin: 0;
     padding: 0;
